Add voice recognition feature with Google Speech Recognition API
This commit adds a VoiceRecognizer module which uses the Google Speech Recognition API to recognize spoken words from the user using their microphone. The language configuration is read from the config file, and a log file is created to record any errors encountered during speech recognition. The config file is also updated to set the language to "ru-RU"